refactor(block-producer): deprecate withdrawal generate, unlock, revert
