+++ crates/tests/src/tests/mod.rs
@@ -7,5 +7,6 @@ mod polyjuice_sender_recover;
 mod restore_mem_block;
 mod restore_mem_pool_pending_withdrawal;
 mod rpc_server;
-mod unlock_withdrawal_to_owner;
 mod finalize_withdrawal_to_owner;
+
+mod deprecated;
